changing all for-loops to size_t

+++ BAM/TCigar.cpp
@@ -16,19 +16,19 @@ namespace BAM {
 // TCigar
 // A class to store, access and manipulate CIGAR operators
 //----------------------------------------------------------
-TCigar::TCigar(TCigar cigar, uint16_t overlapLength, bool isForwardStrand, size_t &mappedBasesClipped) {
-	uint16_t overlap = 0;
+TCigar::TCigar(TCigar cigar, size_t overlapLength, bool isForwardStrand, size_t &mappedBasesClipped) {
+	size_t overlap = 0;
 	//if the overlap is larger than the number of aligned bases in the read (i.e. the other read eclipses this alignment), all bases are softclipped
 	if (overlapLength >= cigar.lengthMapped()) {
 		add('S', cigar.lengthRead());
 	} else {
 		//how many aligned bases before the overlap begins
-		uint16_t nonOverlapLength = cigar.lengthMapped() - overlapLength;
+		size_t nonOverlapLength = cigar.lengthMapped() - overlapLength;
 		//if the read starts before its mate, go beginning->end, otherwise go end->beginning
 		if (isForwardStrand) {
 			std::vector<CigarOperator>::const_iterator iterator = cigar.begin();
 			//copy cigar string until either the start of the overlap or the end of the read is reached
-			int nextLengthMapped = 0;
+			size_t nextLengthMapped = 0;
 			while (lengthMapped() < nonOverlapLength && iterator!=cigar.end()) {
 				//if next segment of cigar string would exceed nonOverlapLength, split it up into M/=/X and S
 				nextLengthMapped = lengthMapped()+iterator->length;
@@ -53,7 +53,7 @@ TCigar::TCigar(TCigar cigar, uint16_t overlapLength, bool isForwardStrand, size_
 		} else {
 			//same as the part above, just use rbegin instead of begin to construct the cigar-string from right to left
 			std::vector<CigarOperator>::const_reverse_iterator iterator = cigar.rbegin();
-			int nextLengthMapped = 0;
+			size_t nextLengthMapped = 0;
 			while (lengthMapped() < nonOverlapLength && iterator != cigar.rend()) {
 				nextLengthMapped = lengthMapped()+iterator->length;
 				if (nextLengthMapped >= nonOverlapLength && (iterator->type == 'D' || iterator->type == 'N')){
@@ -90,7 +90,7 @@ TCigar::TCigar(TCigar cigar, uint16_t overlapLength, bool isForwardStrand, size_
 
 void TCigar::_flipCigar() {
 	std::reverse(_cigar.begin(), _cigar.end());
-	uint32_t temp = lengthSoftClippedLeft();
+	size_t temp = lengthSoftClippedLeft();
 	_lengthSoftClippedLeft = lengthSoftClippedRight();
 	_lengthSoftClippedRight = temp;
 }
@@ -105,7 +105,7 @@ void TCigar::clear() {
 	_lengthSoftClippedRight  = 0;
 };
 
-void TCigar::add(char Type, uint32_t Length) {
+void TCigar::add(char Type, size_t Length) {
 	if (_lengthSoftClippedRight) { UERROR("Cigar string contains entries past soft clipping on right!"); }
 	if (Type == 'M' || Type == '=' || Type == 'X') {
 		_lengthAligned += Length;
