+++ GLF/tests/TTestGLFFile.h
@@ -34,39 +34,39 @@ protected:
 
     //tmp vars for dummy sites
     long _dummyPos; long _dummyDist; long _dummyMaxDist;
-    uint32_t _dummyDepth; uint32_t _dummyMaxDepth;
-    uint8_t _dummy_RMS_mappingQual; uint32_t _dummyMax_RMS_mappingQual;
+    size_t _dummyDepth; size_t _dummyMaxDepth;
+    uint8_t _dummy_RMS_mappingQual; size_t _dummyMax_RMS_mappingQual;
     GenotypeLikelihoods::TGenotypeLikelihoods _dummyGenotypeLikelihoods{1.};
     GenotypeLikelihoods::TGenotypeLikelihoods _dummyGenotypeLikelihoodsHaploid{1.};
 
     void _iteratePosition();
     void _iterateDepth();
     void _iterateRMSMappingQual();
-    virtual void _iterateGenotypeLikelihoods(uint32_t curDepth);
+    virtual void _iterateGenotypeLikelihoods(size_t curDepth);
 
     //storage of written alignments
     std::vector<genometools::TGenomePosition> _writtenPositions;
-    std::vector<uint32_t> _writtenDepths;
+    std::vector<size_t> _writtenDepths;
     std::vector<uint8_t> _writtenRMSMappingQualities;
     std::vector<GenotypeLikelihoods::TGenotypeLikelihoods> _writtenGenotypeLikelihoods;
     std::vector<GenotypeLikelihoods::TGenotypeLikelihoods> _writtenGenotypeLikelihoodsWithMissingSites;
 
-    void _initialize(const std::vector<uint32_t>& ChrLength, const std::vector<uint8_t>& ChrPloidy);
+    void _initialize(const std::vector<size_t>& ChrLength, const std::vector<uint8_t>& ChrPloidy);
 
 public:
     TTestGLFFile(){};
 
-    void openOutput(const std::string & Filename, std::vector<uint32_t>& ChrLength);
-    void openOutput(const std::string & Filename, std::vector<uint32_t>& ChrLength, const std::vector<uint8_t>& ChrPloidy);
+    void openOutput(const std::string & Filename, std::vector<size_t>& ChrLength);
+    void openOutput(const std::string & Filename, std::vector<size_t>& ChrLength, const std::vector<uint8_t>& ChrPloidy);
     void closeOutput();
 
-    void writeSite(long pos, uint32_t depth, GenotypeLikelihoods::TGenotypeLikelihoods &genotypeLikelihoods, uint8_t RMS_mappingQuality);
+    void writeSite(long pos, size_t depth, GenotypeLikelihoods::TGenotypeLikelihoods &genotypeLikelihoods, uint8_t RMS_mappingQuality);
     void writeNewChromosome();
     // write dummy sites with shuffling
-    void writeDummySites(uint32_t numSites);
+    void writeDummySites(size_t numSites);
     void writeDummySite(long pos);
-    void writeDummySite(long pos, uint32_t depth);
-    void writeDummySite(long pos, uint32_t depth, GenotypeLikelihoods::TGenotypeLikelihoods & genotypeLikelihoods);
+    void writeDummySite(long pos, size_t depth);
+    void writeDummySite(long pos, size_t depth, GenotypeLikelihoods::TGenotypeLikelihoods & genotypeLikelihoods);
 
     //getters
     std::string filename()const { return _filename; };
@@ -76,16 +76,16 @@ public:
     std::vector<GenotypeLikelihoods::TGenotypeLikelihoods>::iterator endGenotypeLikelihoods(){ return _writtenGenotypeLikelihoods.end(); };
     std::vector<GenotypeLikelihoods::TGenotypeLikelihoods>::iterator beginGenotypeLikelihoodsWithMissingSites(){ return _writtenGenotypeLikelihoodsWithMissingSites.begin(); };
     std::vector<GenotypeLikelihoods::TGenotypeLikelihoods>::iterator endGenotypeLikelihoodsWithMissingSites(){ return _writtenGenotypeLikelihoodsWithMissingSites.end(); };
-    GenotypeLikelihoods::TGenotypeLikelihoods genotypeLikelihoodsWithMissingSites(uint32_t index){ return _writtenGenotypeLikelihoodsWithMissingSites[index]; };
+    GenotypeLikelihoods::TGenotypeLikelihoods genotypeLikelihoodsWithMissingSites(size_t index){ return _writtenGenotypeLikelihoodsWithMissingSites[index]; };
     std::vector<genometools::TGenomePosition>::iterator beginPositions(){ return _writtenPositions.begin(); };
     std::vector<genometools::TGenomePosition>::iterator endPositions(){ return _writtenPositions.end(); };
-    genometools::TGenomePosition position(uint32_t index){ return _writtenPositions[index]; };
-    std::vector<uint32_t>::iterator beginDepths(){ return _writtenDepths.begin(); };
-    std::vector<uint32_t>::iterator endDepths(){ return _writtenDepths.end(); };
-    uint32_t depth(uint32_t index){ return _writtenDepths[index]; };
+    genometools::TGenomePosition position(size_t index){ return _writtenPositions[index]; };
+    std::vector<size_t>::iterator beginDepths(){ return _writtenDepths.begin(); };
+    std::vector<size_t>::iterator endDepths(){ return _writtenDepths.end(); };
+    size_t depth(size_t index){ return _writtenDepths[index]; };
     std::vector<uint8_t>::iterator beginRMSMappingQualities(){ return _writtenRMSMappingQualities.begin(); };
     std::vector<uint8_t>::iterator endRMSMappingQualities(){ return _writtenRMSMappingQualities.end(); };
-    uint8_t RMSMappingQuality(uint32_t index){ return _writtenRMSMappingQualities[index]; };
+    uint8_t RMSMappingQuality(size_t index){ return _writtenRMSMappingQualities[index]; };
 };
 
 }; // end namespace
